+++ src/lib/state.ts
@@ -1,4 +1,4 @@
-import WireGuard from "@root/reseda";
+import WireGuard from "@components/reseda";
 import { createContext } from "react";
 
 const WireguardContext = createContext<WireGuard>(null);
