+++ tailwind.config.js
@@ -2,7 +2,8 @@ module.exports = {
     content: [
       "./src/pages/**/*.{js,ts,jsx,tsx}",
       "./src/components/*.{js,ts,jsx,tsx}",
-      "./src/components/un-ui/*.{js,ts,jsx,tsx}"
+      "./src/components/un-ui/*.{js,ts,jsx,tsx}",
+      "./src/*.{js,ts,jsx,tsx}"
     ],
     theme: {
         fontFamily: {
